+++ sdk/resourcemanager/containerregistry/armcontainerregistry/fake/importpipelines_server.go
@@ -25,9 +25,9 @@ type ImportPipelinesServer struct {
 	// HTTP status codes to indicate success: http.StatusOK, http.StatusCreated
 	BeginCreate func(ctx context.Context, resourceGroupName string, registryName string, importPipelineName string, importPipelineCreateParameters armcontainerregistry.ImportPipeline, options *armcontainerregistry.ImportPipelinesClientBeginCreateOptions) (resp azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientCreateResponse], errResp azfake.ErrorResponder)
 
-	// BeginDelete is the fake for method ImportPipelinesClient.BeginDelete
-	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted, http.StatusNoContent
-	BeginDelete func(ctx context.Context, resourceGroupName string, registryName string, importPipelineName string, options *armcontainerregistry.ImportPipelinesClientBeginDeleteOptions) (resp azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientDeleteResponse], errResp azfake.ErrorResponder)
+	// Delete is the fake for method ImportPipelinesClient.Delete
+	// HTTP status codes to indicate success: http.StatusOK, http.StatusNoContent
+	Delete func(ctx context.Context, resourceGroupName string, registryName string, importPipelineName string, options *armcontainerregistry.ImportPipelinesClientDeleteOptions) (resp azfake.Responder[armcontainerregistry.ImportPipelinesClientDeleteResponse], errResp azfake.ErrorResponder)
 
 	// Get is the fake for method ImportPipelinesClient.Get
 	// HTTP status codes to indicate success: http.StatusOK
@@ -45,7 +45,6 @@ func NewImportPipelinesServerTransport(srv *ImportPipelinesServer) *ImportPipeli
 	return &ImportPipelinesServerTransport{
 		srv:          srv,
 		beginCreate:  newTracker[azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientCreateResponse]](),
-		beginDelete:  newTracker[azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientDeleteResponse]](),
 		newListPager: newTracker[azfake.PagerResponder[armcontainerregistry.ImportPipelinesClientListResponse]](),
 	}
 }
@@ -55,7 +54,6 @@ func NewImportPipelinesServerTransport(srv *ImportPipelinesServer) *ImportPipeli
 type ImportPipelinesServerTransport struct {
 	srv          *ImportPipelinesServer
 	beginCreate  *tracker[azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientCreateResponse]]
-	beginDelete  *tracker[azfake.PollerResponder[armcontainerregistry.ImportPipelinesClientDeleteResponse]]
 	newListPager *tracker[azfake.PagerResponder[armcontainerregistry.ImportPipelinesClientListResponse]]
 }
 
@@ -84,8 +82,8 @@ func (i *ImportPipelinesServerTransport) dispatchToMethodFake(req *http.Request,
 			switch method {
 			case "ImportPipelinesClient.BeginCreate":
 				res.resp, res.err = i.dispatchBeginCreate(req)
-			case "ImportPipelinesClient.BeginDelete":
-				res.resp, res.err = i.dispatchBeginDelete(req)
+			case "ImportPipelinesClient.Delete":
+				res.resp, res.err = i.dispatchDelete(req)
 			case "ImportPipelinesClient.Get":
 				res.resp, res.err = i.dispatchGet(req)
 			case "ImportPipelinesClient.NewListPager":
@@ -161,51 +159,40 @@ func (i *ImportPipelinesServerTransport) dispatchBeginCreate(req *http.Request) 
 	return resp, nil
 }
 
-func (i *ImportPipelinesServerTransport) dispatchBeginDelete(req *http.Request) (*http.Response, error) {
-	if i.srv.BeginDelete == nil {
-		return nil, &nonRetriableError{errors.New("fake for method BeginDelete not implemented")}
+func (i *ImportPipelinesServerTransport) dispatchDelete(req *http.Request) (*http.Response, error) {
+	if i.srv.Delete == nil {
+		return nil, &nonRetriableError{errors.New("fake for method Delete not implemented")}
 	}
-	beginDelete := i.beginDelete.get(req)
-	if beginDelete == nil {
-		const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.ContainerRegistry/registries/(?P<registryName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/importPipelines/(?P<importPipelineName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
-		regex := regexp.MustCompile(regexStr)
-		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if matches == nil || len(matches) < 4 {
-			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
-		}
-		resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
-		if err != nil {
-			return nil, err
-		}
-		registryNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("registryName")])
-		if err != nil {
-			return nil, err
-		}
-		importPipelineNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("importPipelineName")])
-		if err != nil {
-			return nil, err
-		}
-		respr, errRespr := i.srv.BeginDelete(req.Context(), resourceGroupNameParam, registryNameParam, importPipelineNameParam, nil)
-		if respErr := server.GetError(errRespr, req); respErr != nil {
-			return nil, respErr
-		}
-		beginDelete = &respr
-		i.beginDelete.add(req, beginDelete)
+	const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.ContainerRegistry/registries/(?P<registryName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/importPipelines/(?P<importPipelineName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
+	regex := regexp.MustCompile(regexStr)
+	matches := regex.FindStringSubmatch(req.URL.EscapedPath())
+	if matches == nil || len(matches) < 4 {
+		return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
 	}
-
-	resp, err := server.PollerResponderNext(beginDelete, req)
+	resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
 	if err != nil {
 		return nil, err
 	}
-
-	if !contains([]int{http.StatusOK, http.StatusAccepted, http.StatusNoContent}, resp.StatusCode) {
-		i.beginDelete.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted, http.StatusNoContent", resp.StatusCode)}
+	registryNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("registryName")])
+	if err != nil {
+		return nil, err
 	}
-	if !server.PollerResponderMore(beginDelete) {
-		i.beginDelete.remove(req)
+	importPipelineNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("importPipelineName")])
+	if err != nil {
+		return nil, err
 	}
-
+	respr, errRespr := i.srv.Delete(req.Context(), resourceGroupNameParam, registryNameParam, importPipelineNameParam, nil)
+	if respErr := server.GetError(errRespr, req); respErr != nil {
+		return nil, respErr
+	}
+	respContent := server.GetResponseContent(respr)
+	if !contains([]int{http.StatusOK, http.StatusNoContent}, respContent.HTTPStatus) {
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusNoContent", respContent.HTTPStatus)}
+	}
+	resp, err := server.NewResponse(respContent, req, nil)
+	if err != nil {
+		return nil, err
+	}
 	return resp, nil
 }